fix: 일기 초기화 버튼 삭제 & 띠로리 앨범 버튼으로 교체
