--- src/components/diary/image/ImagesHistoryButton.tsx
@@ -1,39 +1,0 @@
-import React from "react";
-
-interface Props {
-  onClick?: () => void;
-  isFull?: boolean;
-}
-
-const ImagesHistoryButton: React.FC<Props> = ({ onClick, isFull }) => {
-  return (
-    <button onClick={onClick}>
-      <svg
-        width="52"
-        height="52"
-        viewBox="0 0 52 52"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <path
-          fill-rule="evenodd"
-          clip-rule="evenodd"
-          d="M6.5 9.75C6.5 7.95507 7.95507 6.5 9.75 6.5H42.25C44.0449 6.5 45.5 7.95507 45.5 9.75V42.25C45.5 44.0449 44.0449 45.5 42.25 45.5H9.75C7.95507 45.5 6.5 44.0449 6.5 42.25V9.75ZM42.25 9.75H9.75V42.25H42.25V9.75Z"
-          fill={`${isFull ? "#F46666" : "#000000"}`}
-        />
-        <path
-          fill-rule="evenodd"
-          clip-rule="evenodd"
-          d="M32.8698 22.0502C33.2673 21.8837 33.694 21.798 34.125 21.798C34.556 21.798 34.9827 21.8837 35.3802 22.0502C35.7754 22.2157 36.134 22.4577 36.4353 22.7622L45.024 31.351C45.6586 31.9856 45.6586 33.0144 45.024 33.649C44.3894 34.2837 43.3606 34.2837 42.726 33.649L34.125 25.0481L34.1163 25.0569L25.0605 34.1126C24.7591 34.4172 24.4005 34.6593 24.0052 34.8248C23.6077 34.9912 23.181 35.077 22.75 35.077C22.319 35.077 21.8923 34.9912 21.4948 34.8248C21.0995 34.6593 20.7408 34.4172 20.4394 34.1125L16.25 29.9231L16.2413 29.9319L9.27405 36.899C8.63945 37.5337 7.61055 37.5337 6.97595 36.899C6.34135 36.2644 6.34135 35.2356 6.97595 34.601L13.9395 27.6374C14.2409 27.3328 14.5995 27.0907 14.9948 26.9252C15.3923 26.7588 15.819 26.673 16.25 26.673C16.681 26.673 17.1077 26.7588 17.5052 26.9252C17.9005 27.0907 18.2592 27.3328 18.5605 27.6374L22.75 31.8269L22.7587 31.8181L31.8145 22.7624C32.1159 22.4578 32.4745 22.2157 32.8698 22.0502Z"
-          fill={`${isFull ? "#F46666" : "#000000"}`}
-        />
-        <path
-          d="M20.3125 21.125C21.6587 21.125 22.75 20.0337 22.75 18.6875C22.75 17.3413 21.6587 16.25 20.3125 16.25C18.9663 16.25 17.875 17.3413 17.875 18.6875C17.875 20.0337 18.9663 21.125 20.3125 21.125Z"
-          fill={`${isFull ? "#F46666" : "#000000"}`}
-        />
-      </svg>
-    </button>
-  );
-};
-
-export default ImagesHistoryButton;
